rug_pull_monitor/src/requirements.txt
requests==2.32.3
requests-cache==1.2.1
numpy==1.26.4
python-dotenv==0.21.0
schedule==1.2.2